requirements.txt
streamlit>=1.33.0
pymupdf>=1.24.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
pandas>=2.0.0
requests>=2.31.0
numpy>=1.26.0
python-docx>=1.1.0
reportlab>=4.0.0

# Development dependencies
pytest>=7.0.0
black>=24.0.0
flake8>=7.0.0
mypy>=1.0.0
pdfplumber>=0.11.0
